Set Up Cypress
Cypress is an versatile testing framework that can be used for unit and
end-to-end frontend tests. It utilizes mocha and chai libraries for
writing tests.

The flow libdefs for all of these libraries are quite lacking, however
and I'm strongly considering using a .flow file that is localized to the
cypress folder to resolve this, and additionally, mitigate collisions
with Jest.

Cypress has not been hooked up to the larger testing sequence we have,
	as no useful test have been written yet.

test plan: ensure all existing automated test pass