Implement PayableAddressStore for Ledger Accounts
PayableAddressStore maps addresses of EVM-based currencies to a participant's
address capable of accepting the currency. This structure exists to
accomodate safe migration for grain/payout token changes. Users must verify
themselves that the address they are supplying is capable of receiving their
share of a grain distribution.

An address may be deleted by passing in an empty string for the
`address` parameter. This is useful in case the underlying private key
is compromised or the exchange hosting a custodial account is hacked.

test plan: Unit test are provided